Select and play the audio element corresponding to key pressed
The audio elements are tagged with a 'data-key' attribute
which matches the value of the 'keyCode' property of the
keydown event. For example, 65 for 'a'.

We select the matching audio element, if any, with a call
to document.querySelector(), which returns the first matching
element for given CSS selector. We build the CSS selector
dynamically, to insert the keyCode in the following format:

  audio[data-key="keyCode"]

e.g.

  audio[data-key="65"]

The quotes around the value are required here, otherwise the
search results in an error.

In order to limit the number of searches in the DOM, I added a
preliminary check on the value of the keyCode, to verify that it
is within the expected boundaries (65..83) before looking for the
corresponding audio element.

Not all keyCode values in this range have a matching audio element
though; a second check is thus necessary to skip 'null' values
returned by document.querySelector() when no match is found.

Finally, we play the audio by calling the play() method of the
audio element.

We notice that hitting the same key multiple times before the
sound currently playing has ended has no effect. A call to play()
on an audio element already playing is ignored.