Display visual feedback when a key is pressed
At this step, the visual feedback never disappears: we are adding the
CSS class 'playing' to trigger the feedback, but we are never removing
the class.

I also refactored the code to play the audio element, now checking that
the element is not null and calling the code, instead of checking when
the element is null and returning. In this way, I can check separately
for the presence of the visual key element and the presence of the
corresponding audio element. If only one is present, the expected effect
will trigger, only on that element.