Apply a short transition to the hands, with a bounce
Instead of moving instantly from on position to the next,
the hands now travel over a very short duration (0.05s)
and bounce due to the peculiar shape of the cubic bezier
curve used as transition-timing-function (this function
can be seen and customized in the Web Developer Tools,
by clicking the icon next to the value in the CSS inspector).

The transition can be tested interactively in the Web Developer
tools, by changing the angle value in the rotate() transform.