Wrap our code in a closure to make our future variables private
I do not want to define global variables which could interact
with other browser components in an unexpected way. I wrapped
the code in a function, and the variables that will be defined
within will be private to this function.

I defined a utility function, privately(), which makes the usage
of the wrapping function to define a private scope explicit.