Select and rotate each hand according to computed angle
I defined a rotate() function which rotates a hand element
by applying a rotate() transform to its style.transform property.

We notice that the time is off: the computed value replaces
the previous rotate(90deg) transform which set the hands upright.
This initial angle shall not be set in CSS, but used as an offset
in JavaScript.

As a result, the rotation applied will no longer be between 0 and 360
degrees, but between 90 and 450 degrees (which works as expected).
The rotation was actually not between 0 and 360 for hours already,
but between 0 and 720 because Date#getHours() returns a value between
0 and 23. The rotation for hours is now between 90 and 810 degrees.