Remove the transition
I will move all the hands smoothly instead, e.g. 10 times per second.
This will make all the hands move all the time, which is more accurate
and more realistic: the hour hand does not stay in the same place for
a whole hour on a real clock: it moves progressively to the next hour.